Refactor convention membership helpers
