+++ src/core/code_audit/conventions.rs
@@ -7,21 +7,15 @@
 use std::collections::HashMap;
 use std::path::Path;
 
+use super::convention_membership::{
+    declared_trait_name, declares_type_subject, is_convention_exception, is_utility_like_file,
+    member_requirement_deviation,
+};
 use super::fingerprint::FileFingerprint;
 use super::import_matching::has_import_with_context;
 use super::naming::{detect_naming_suffix, suffix_matches};
 use super::signatures::{compute_signature_skeleton, tokenize_signature};
 use crate::core::component::AuditConfig;
-
-const GENERIC_UTILITY_SUFFIXES: &[&str] = &[
-    "Base",
-    "Handler",
-    "Handlers",
-    "Helper",
-    "Helpers",
-    "Projector",
-    "Projectors",
-];
 
 #[derive(Debug, Clone, Default, PartialEq, Eq, Hash, serde::Serialize, serde::Deserialize)]
 #[serde(rename_all = "lowercase")]
@@ -690,62 +684,6 @@ pub fn discover_conventions_with_config(
         total_files: total,
         confidence,
     })
-}
-
-fn member_requirement_deviation(
-    kind: AuditFinding,
-    description_label: &str,
-    suggestion_verb: &str,
-    expected: &str,
-    expected_suffix: &str,
-    group_name: &str,
-) -> Deviation {
-    Deviation {
-        kind,
-        description: format!("{}: {}", description_label, expected),
-        suggestion: format!(
-            "{} {}{} to match the convention in {}",
-            suggestion_verb, expected, expected_suffix, group_name
-        ),
-    }
-}
-
-fn declared_trait_name(fp: &FileFingerprint) -> Option<String> {
-    let re = regex::Regex::new(r"(?m)^\s*trait\s+([A-Za-z_][A-Za-z0-9_]*)\b").ok()?;
-    re.captures(&fp.content)
-        .and_then(|cap| cap.get(1))
-        .map(|m| m.as_str().to_string())
-}
-
-fn declares_type_subject(fp: &FileFingerprint) -> bool {
-    fp.type_name.is_some() || !fp.type_names.is_empty()
-}
-
-fn is_utility_like_file(fp: &FileFingerprint, audit_config: &AuditConfig) -> bool {
-    let names_to_check: Vec<&str> = if !fp.type_names.is_empty() {
-        fp.type_names.iter().map(|s| s.as_str()).collect()
-    } else {
-        fp.type_name.as_deref().into_iter().collect()
-    };
-
-    declared_trait_name(fp).is_some()
-        || names_to_check.iter().any(|name| {
-            GENERIC_UTILITY_SUFFIXES
-                .iter()
-                .any(|suffix| name.ends_with(suffix))
-                || audit_config
-                    .utility_suffixes
-                    .iter()
-                    .any(|suffix| name.ends_with(suffix))
-        })
-}
-
-fn is_convention_exception(fp: &FileFingerprint, audit_config: &AuditConfig) -> bool {
-    let normalized = fp.relative_path.replace('\\', "/");
-    audit_config
-        .convention_exception_globs
-        .iter()
-        .any(|pattern| glob_match::glob_match(pattern, &normalized))
 }
 
 // ============================================================================
