+++ src/core/code_audit/mod.rs
@@ -18,6 +18,7 @@ mod comment_blocks;
 mod comment_hygiene;
 pub mod compare;
 mod compiler_warnings;
+mod convention_membership;
 pub(crate) mod conventions;
 pub(crate) mod core_fingerprint;
 mod dead_code;
